refactor: change jarust_transport file structure
