+++ jarust/examples/raw_echotest.rs
@@ -3,7 +3,7 @@ use jarust::jaconfig::JaConfig;
 use jarust::jaconnection::CreateConnectionParams;
 use jarust::japlugin::AttachHandleParams;
 use jarust::prelude::Attach;
-use jarust_transport::transaction_gen::RandomTransactionGenerator;
+use jarust_transport::tgenerator::RandomTransactionGenerator;
 use serde_json::json;
 use std::time::Duration;
 use tracing_subscriber::EnvFilter;
