+++ jarust/tests/mocks/mock_interface.rs
@@ -6,8 +6,8 @@ use jarust_transport::handle_msg::HandleMessage;
 use jarust_transport::handle_msg::HandleMessageWithEstablishment;
 use jarust_transport::handle_msg::HandleMessageWithEstablishmentAndTimeout;
 use jarust_transport::handle_msg::HandleMessageWithTimeout;
-use jarust_transport::interface::janus_interface::ConnectionParams;
-use jarust_transport::interface::janus_interface::JanusInterface;
+use jarust_transport::janus_interface::ConnectionParams;
+use jarust_transport::janus_interface::JanusInterface;
 use jarust_transport::japrotocol::JaSuccessProtocol;
 use jarust_transport::japrotocol::ResponseType;
 use jarust_transport::prelude::JaTransportResult;
