+++ jarust_plugins/examples/echotest_start_with_est.rs
@@ -8,7 +8,7 @@ use jarust_plugins::echo_test::msg_options::StartOptions;
 use jarust_transport::japrotocol::EstablishmentProtocol;
 use jarust_transport::japrotocol::Jsep;
 use jarust_transport::japrotocol::JsepType;
-use jarust_transport::transaction_gen::RandomTransactionGenerator;
+use jarust_transport::tgenerator::RandomTransactionGenerator;
 use std::path::Path;
 use std::time::Duration;
 use tracing_subscriber::EnvFilter;
