--- jarust_transport/src/interface/mod.rs
@@ -1,5 +1,0 @@
-pub mod janus_interface;
-pub mod restful_interface;
-pub mod websocket_interface;
-
-mod websocket_client;
