+++ jarust_transport/src/restful/mod.rs
@@ -1,0 +1,3 @@
+pub mod restful_interface;
+
+pub use restful_interface::RestfulInterface;
