+++ jarust_transport/src/restful/restful_interface.rs
@@ -3,16 +3,16 @@ use crate::handle_msg::HandleMessage;
 use crate::handle_msg::HandleMessageWithEstablishment;
 use crate::handle_msg::HandleMessageWithEstablishmentAndTimeout;
 use crate::handle_msg::HandleMessageWithTimeout;
-use crate::interface::janus_interface::ConnectionParams;
-use crate::interface::janus_interface::JanusInterface;
+use crate::janus_interface::ConnectionParams;
+use crate::janus_interface::JanusInterface;
 use crate::japrotocol::EstablishmentProtocol;
 use crate::japrotocol::JaResponse;
 use crate::japrotocol::JaSuccessProtocol;
 use crate::japrotocol::ResponseType;
 use crate::prelude::JaTransportResult;
 use crate::respones::ServerInfoRsp;
-use crate::transaction_gen::GenerateTransaction;
-use crate::transaction_gen::TransactionGenerator;
+use crate::tgenerator::GenerateTransaction;
+use crate::tgenerator::TransactionGenerator;
 use jarust_rt::JaTask;
 use serde_json::json;
 use serde_json::Value;
