+++ jarust_transport/src/websocket/router.rs
@@ -1,5 +1,5 @@
-use super::japrotocol::JaResponse;
 use crate::error::JaTransportError;
+use crate::japrotocol::JaResponse;
 use crate::prelude::JaTransportResult;
 use serde_json::Value;
 use std::collections::HashMap;
@@ -131,9 +131,9 @@ impl Router {
 
 #[cfg(test)]
 mod tests {
-    use super::super::japrotocol::JaResponse;
-    use super::super::japrotocol::ResponseType;
     use super::Router;
+    use crate::japrotocol::JaResponse;
+    use crate::japrotocol::ResponseType;
 
     #[tokio::test]
     async fn test_basic_usage() {
